requirements.txt
asgiref==3.5.2
Django==4.0.4
numpy==1.24.2
Pillow==9.5.0
protobuf==4.21.5
six==1.16.0
sklearn==0.0.post4
sqlparse==0.4.2
